requirements.txt
pika==0.11.2
pkg-resources==0.0.0
PyYAML==3.12
